冗長なdir削除しました: build_structured_information
